+++ src/API.jl
@@ -138,15 +138,15 @@ function require_not_empty(pkgs, f::Symbol)
 end
 
 # Provide some convenience calls
-for f in (:develop, :add, :rm, :up, :pin, :free, :test, :build, :status, :why)
+for f in (:develop, :add, :rm, :up, :pin, :free, :test, :build, :status, :why, :precompile)
     @eval begin
         $f(pkg::Union{AbstractString, PackageSpec}; kwargs...) = $f([pkg]; kwargs...)
         $f(pkgs::Vector{<:AbstractString}; kwargs...)          = $f([PackageSpec(pkg) for pkg in pkgs]; kwargs...)
         function $f(pkgs::Vector{PackageSpec}; io::IO=$(f === :status ? :stdout_f : :stderr_f)(), kwargs...)
-            Registry.download_default_registries(io)
+            $(f != :precompile) && Registry.download_default_registries(io)
             ctx = Context()
             # Save initial environment for undo/redo functionality
-            if !saved_initial_snapshot[]
+            if $(f != :precompile) && !saved_initial_snapshot[]
                 add_snapshot_to_undo(ctx.env)
                 saved_initial_snapshot[] = true
             end
@@ -327,6 +327,10 @@ function up(ctx::Context, pkgs::Vector{PackageSpec};
             skip_writing_project::Bool=false,
             kwargs...)
     Context!(ctx; kwargs...)
+    if Operations.is_fully_pinned(ctx)
+        printpkgstyle(ctx.io, :Update, "All dependencies are pinned - nothing to update.", color = Base.info_color())
+        return
+    end
     if update_registry
         Registry.download_default_registries(ctx.io)
         Operations.update_registries(ctx; force=true)
@@ -355,7 +359,7 @@ function pin(ctx::Context, pkgs::Vector{PackageSpec}; all_pkgs::Bool=false, kwar
     Context!(ctx; kwargs...)
     if all_pkgs
         !isempty(pkgs) && pkgerror("cannot specify packages when operating on all packages")
-        append_all_pkgs!(pkgs, ctx, PKGMODE_PROJECT)
+        append_all_pkgs!(pkgs, ctx, PKGMODE_MANIFEST)
     else
         require_not_empty(pkgs, :pin)
     end
@@ -390,7 +394,7 @@ function free(ctx::Context, pkgs::Vector{PackageSpec}; all_pkgs::Bool=false, kwa
     Context!(ctx; kwargs...)
     if all_pkgs
         !isempty(pkgs) && pkgerror("cannot specify packages when operating on all packages")
-        append_all_pkgs!(pkgs, ctx, PKGMODE_PROJECT)
+        append_all_pkgs!(pkgs, ctx, PKGMODE_MANIFEST)
     else
         require_not_empty(pkgs, :free)
     end
@@ -869,10 +873,18 @@ function gc(ctx::Context=Context(); collect_delay::Period=Day(7), verbose=false,
                 uuid_dir = joinpath(scratchdir, uuid)
                 !isdir(uuid_dir) && continue
                 for space in readdir(uuid_dir)
-                    space_dir = joinpath(uuid_dir, space)
-                    !isdir(space_dir) && continue
-                    if !(space_dir in spaces_to_keep)
-                        push!(depot_orphaned_scratchspaces, space_dir)
+                    space_dir_or_file = joinpath(uuid_dir, space)
+                    if isdir(space_dir_or_file)
+                        if !(space_dir_or_file in spaces_to_keep)
+                            push!(depot_orphaned_scratchspaces, space_dir_or_file)
+                        end
+                    elseif uuid == Operations.PkgUUID && isfile(space_dir_or_file)
+                        # special cleanup for the precompile cache files that Pkg saves
+                        if any(prefix->startswith(basename(space_dir_or_file), prefix), ("suspend_cache_", "pending_cache_"))
+                            if mtime(space_dir_or_file) < (time() - (24*60*60))
+                                push!(depot_orphaned_scratchspaces, space_dir_or_file)
+                            end
+                        end
                     end
                 end
             end
@@ -928,7 +940,16 @@ function gc(ctx::Context=Context(); collect_delay::Period=Day(7), verbose=false,
 
     # Delete paths for unreachable package versions and artifacts, and computing size saved
     function delete_path(path)
-        path_size = recursive_dir_size(path)
+        path_size = if isfile(path)
+            try
+                lstat(path).size
+            catch ex
+                @error("Failed to calculate size of $path", exception=ex)
+                0
+            end
+        else
+            recursive_dir_size(path)
+        end
         try
             Base.Filesystem.prepare_for_deletion(path)
             Base.rm(path; recursive=true, force=true)
@@ -1032,39 +1053,29 @@ function build(ctx::Context, pkgs::Vector{PackageSpec}; verbose=false, kwargs...
     Operations.build(ctx, Set{UUID}(pkg.uuid for pkg in pkgs), verbose)
 end
 
+# should sync with the types of arguments of `Base.stale_cachefile`
+const StaleCacheKey = Tuple{Base.PkgId, UInt128, String, String}
 
-function _is_stale!(stale_cache::Dict, paths::Vector{String}, sourcepath::String)
+function _is_stale!(stale_cache::Dict{StaleCacheKey,Bool}, paths::Vector{String}, sourcepath::String)
     for path_to_try in paths
         staledeps = Base.stale_cachefile(sourcepath, path_to_try, ignore_loaded = true)
         if staledeps === true
             continue
         end
-        staledeps, ocachefile = staledeps::Tuple{Vector{Any}, Union{Nothing, String}}
+        staledeps, _ = staledeps::Tuple{Vector{Any}, Union{Nothing, String}}
         # finish checking staledeps module graph
         for i in 1:length(staledeps)
             dep = staledeps[i]
             dep isa Module && continue
             modpath, modkey, modbuild_id = dep::Tuple{String, Base.PkgId, UInt128}
             modpaths = Base.find_all_in_cache_path(modkey)
-            modfound = false
             for modpath_to_try in modpaths::Vector{String}
-                modstaledeps = get!(() -> Base.stale_cachefile(modkey, modbuild_id, modpath, modpath_to_try), 
-                                    stale_cache, (modkey, modbuild_id, modpath, modpath_to_try))
-                if modstaledeps === true
-                    continue
-                end
-                modstaledeps, modocachepath = modstaledeps::Tuple{Vector{Any}, Union{Nothing, String}}
-                staledeps[i] = (modpath, modkey, modpath_to_try, modstaledeps, modocachepath)
-                modfound = true
-                break
+                stale_cache_key = (modkey, modbuild_id, modpath, modpath_to_try)::StaleCacheKey
+                is_stale = get!(() -> Base.stale_cachefile(stale_cache_key...) === true,
+                                stale_cache, stale_cache_key)
+                is_stale && continue
+                @goto check_next_path
             end
-            if !modfound
-                staledeps = true
-                break
-            end
-        end
-        if staledeps === true
-            continue
         end
         try
             # update timestamp of precompilation file so that it is the first to be tried by code loading
@@ -1074,6 +1085,7 @@ function _is_stale!(stale_cache::Dict, paths::Vector{String}, sourcepath::String
             ex isa Base.IOError || rethrow()
         end
         return false
+        @label check_next_path
     end
     return true
 end
@@ -1097,12 +1109,8 @@ function get_or_make_pkgspec(pkgspecs::Vector{PackageSpec}, ctx::Context, uuid)
     end
 end
 
-precompile(pkgs...; kwargs...) = precompile(Context(), [pkgs...]; kwargs...)
-precompile(pkg::String; kwargs...) = precompile(Context(), pkg; kwargs...)
-precompile(ctx::Context, pkg::String; kwargs...) = precompile(ctx, [pkg]; kwargs...)
-precompile(pkgs::Vector{String}=String[]; kwargs...) = precompile(Context(), pkgs; kwargs...)
-function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::Bool=false,
-                    strict::Bool=false, warn_loaded = true, already_instantiated = false, kwargs...)
+function precompile(ctx::Context, pkgs::Vector{PackageSpec}; internal_call::Bool=false,
+                    strict::Bool=false, warn_loaded = true, already_instantiated = false, timing::Bool = false, kwargs...)
     Context!(ctx; kwargs...)
     already_instantiated || instantiate(ctx; allow_autoprecomp=false, kwargs...)
     time_start = time_ns()
@@ -1115,7 +1123,7 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
     num_tasks = parse(Int, get(ENV, "JULIA_NUM_PRECOMPILE_TASKS", string(default_num_tasks)))
     parallel_limiter = Base.Semaphore(num_tasks)
     io = ctx.io
-    fancyprint = can_fancyprint(io)
+    fancyprint = can_fancyprint(io) && !timing
 
     recall_precompile_state() # recall suspended and force-queued packages
     !internal_call && precomp_unsuspend!() # when manually called, unsuspend all packages that were suspended due to precomp errors
@@ -1124,7 +1132,7 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
         Base.PkgId(uuid, name)
         for (name, uuid) in ctx.env.project.deps if !Base.in_sysimage(Base.PkgId(uuid, name))
     ]
-    stale_cache = Dict{Tuple{Base.PkgId, UInt128, String, String}, Union{Tuple{Vector{Any}, String}, Bool}}()
+    stale_cache = Dict{StaleCacheKey, Bool}()
     exts = Dict{Base.PkgId, String}() # ext -> parent
     # make a flat map of each dep and its deps
     depsmap = Dict{Base.PkgId, Vector{Base.PkgId}}()
@@ -1206,6 +1214,7 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
 
     # if a list of packages is given, restrict to dependencies of given packages
     if !isempty(pkgs)
+        pkgs_names = [p.name for p in pkgs]
         function collect_all_deps(depsmap, dep, alldeps=Base.PkgId[])
             append!(alldeps, depsmap[dep])
             for _dep in depsmap[dep]
@@ -1215,15 +1224,17 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
         end
         keep = Base.PkgId[]
         for dep in depsmap
-            if first(dep).name in pkgs
+            if first(dep).name in pkgs_names
                 push!(keep, first(dep))
                 append!(keep, collect_all_deps(depsmap, first(dep)))
             end
         end
         filter!(d->in(first(d), keep), depsmap)
-        isempty(depsmap) && pkgerror("No direct dependencies found matching $(repr(pkgs))")
+        isempty(depsmap) && pkgerror("No direct dependencies found matching $(repr(pkgs_names))")
+        target = join(pkgs_names, ", ")
+    else
+        target = "project..."
     end
-    target = string(isempty(pkgs) ? "project" : join(pkgs, ", "), "...")
 
     pkg_queue = Base.PkgId[]
     failed_deps = Dict{Base.PkgId, String}()
@@ -1248,14 +1259,17 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
     n_done::Int = 0
     n_already_precomp::Int = 0
     n_loaded::Int = 0
+    interrupted = false
 
-    function handle_interrupt(err)
+    function handle_interrupt(err, in_printloop = false)
         notify(interrupted_or_done)
         sleep(0.2) # yield for a period to let the print loop cease first
+        in_printloop || wait(t_print) # wait to let the print loop cease first
         if err isa InterruptException
             lock(print_lock) do
                 println(io, " Interrupted: Exiting precompilation...")
             end
+            interrupted = true
         else
             @error "Pkg.precompile error" exception=(err, catch_backtrace())
         end
@@ -1331,11 +1345,13 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
                 wait(t)
             end
         catch err
-            handle_interrupt(err)
+            handle_interrupt(err, true)
         finally
             fancyprint && print(io, ansi_enablecursor)
         end
     end
+    stderr_buffers = Dict{Base.PkgId,IOBuffer}()
+    stderr_outputs = Dict{Base.PkgId,String}()
     tasks = Task[]
     Base.LOADING_CACHE[] = Base.LoadingCache()
     ## precompilation loop
@@ -1372,6 +1388,7 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
                     Base.acquire(parallel_limiter)
                     is_direct_dep = pkg in direct_deps
                     iob = IOBuffer()
+                    stderr_buffers[pkg] = iob
                     _name = haskey(exts, pkg) ? string(exts[pkg], " → ", pkg.name) : pkg.name
                     name = is_direct_dep ? _name : string(color_string(_name, :light_black))
                     !fancyprint && lock(print_lock) do
@@ -1386,23 +1403,24 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
                         return
                     end
                     try
-                        ret = Logging.with_logger(Logging.NullLogger()) do
+                        t = @elapsed ret = Logging.with_logger(Logging.NullLogger()) do
                             # TODO: Explore allowing parallel LLVM image generation. Needs careful load balancing
                             withenv("JULIA_IMAGE_THREADS" => "1") do
                                 # capture stderr, send stdout to devnull, don't skip loaded modules
                                 Base.compilecache(pkg, sourcepath, iob, devnull, false)
                             end
                         end
+                        t_str = timing ? string(lpad(round(t * 1e3, digits = 1), 9), " ms") : ""
                         if ret isa Base.PrecompilableError
                             push!(precomperr_deps, pkg)
                             precomp_queue!(get_or_make_pkgspec(pkg_specs, ctx, pkg.uuid))
                             !fancyprint && lock(print_lock) do
-                                println(io, string(color_string("  ? ", Base.warn_color()), name))
+                                println(io, t_str, color_string("  ? ", Base.warn_color()), name)
                             end
                         else
                             queued && precomp_dequeue!(get_or_make_pkgspec(pkg_specs, ctx, pkg.uuid))
                             !fancyprint && lock(print_lock) do
-                                println(io, string(color_string("  ✓ ", loaded ? Base.warn_color() : :green), name))
+                                println(io, t_str, color_string("  ✓ ", loaded ? Base.warn_color() : :green), name)
                             end
                             was_recompiled[pkg] = true
                         end
@@ -1410,14 +1428,22 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
                     catch err
                         if err isa ErrorException || (err isa ArgumentError && startswith(err.msg, "Invalid header in cache file"))
                             failed_deps[pkg] = (strict || is_direct_dep) ? string(sprint(showerror, err), "\n", String(take!(iob))) : ""
+                            delete!(stderr_buffers, pkg)
                             !fancyprint && lock(print_lock) do
-                                println(io, string(color_string("  ✗ ", Base.error_color()), name))
+                                println(io, timing ? " "^9 : "", color_string("  ✗ ", Base.error_color()), name)
                             end
                             queued && precomp_dequeue!(get_or_make_pkgspec(pkg_specs, ctx, pkg.uuid))
                             precomp_suspend!(get_or_make_pkgspec(pkg_specs, ctx, pkg.uuid))
                         else
-                            rethrow(err)
+                            rethrow()
                         end
+                    else
+                        # otherwise capture any stderr output as warnings
+                        out = strip(String(take!(iob)))
+                        if !isempty(out)
+                            stderr_outputs[pkg] = out
+                        end
+                        delete!(stderr_buffers, pkg)
                     finally
                         Base.release(parallel_limiter)
                     end
@@ -1447,41 +1473,62 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
     end
     notify(first_started) # in cases of no-op or !fancyprint
     save_precompile_state() # save lists to scratch space
-    wait(t_print)
-    !all(istaskdone, tasks) && return # if some not finished, must have errored or been interrupted
+    fancyprint && wait(t_print)
+    quick_exit = !all(istaskdone, tasks) || interrupted # if some not finished internal error is likely
+    if quick_exit # collect any stderr output from any unfinished precompile tasks
+        for (pkg, iob) in stderr_buffers
+            out = strip(String(take!(iob)))
+            if !isempty(out)
+                stderr_outputs[pkg] = out
+            end
+        end
+    end
     seconds_elapsed = round(Int, (time_ns() - time_start) / 1e9)
     ndeps = count(values(was_recompiled))
-    if ndeps > 0 || !isempty(failed_deps)
-        plural = ndeps == 1 ? "y" : "ies"
+    if ndeps > 0 || !isempty(failed_deps) || (quick_exit && !isempty(stderr_outputs))
         str = sprint() do iostr
-            print(iostr, "  $(ndeps) dependenc$(plural) successfully precompiled in $(seconds_elapsed) seconds")
-            if n_already_precomp > 0 || !isempty(circular_deps) || !isempty(skipped_deps)
-                n_already_precomp > 0 && (print(iostr, ". $n_already_precomp already precompiled"))
-                !isempty(circular_deps) && (print(iostr, ". $(length(circular_deps)) skipped due to circular dependency"))
-                !isempty(skipped_deps) && (print(iostr, ". $(length(skipped_deps)) skipped during auto due to previous errors"))
-                print(iostr, ".")
+            if !quick_exit
+                plural = ndeps == 1 ? "y" : "ies"
+                print(iostr, "  $(ndeps) dependenc$(plural) successfully precompiled in $(seconds_elapsed) seconds")
+                if n_already_precomp > 0 || !isempty(circular_deps) || !isempty(skipped_deps)
+                    n_already_precomp > 0 && (print(iostr, ". $n_already_precomp already precompiled"))
+                    !isempty(circular_deps) && (print(iostr, ". $(length(circular_deps)) skipped due to circular dependency"))
+                    !isempty(skipped_deps) && (print(iostr, ". $(length(skipped_deps)) skipped during auto due to previous errors"))
+                    print(iostr, ".")
+                end
+                if n_loaded > 0
+                    plural1 = n_loaded == 1 ? "y" : "ies"
+                    plural2 = n_loaded == 1 ? "a different version is" : "different versions are"
+                    plural3 = n_loaded == 1 ? "" : "s"
+                    print(iostr, "\n  ",
+                        color_string(string(n_loaded), Base.warn_color()),
+                        " dependenc$(plural1) precompiled but $(plural2) currently loaded. Restart julia to access the new version$(plural3)"
+                    )
+                end
+                if !isempty(precomperr_deps)
+                    pluralpc = length(precomperr_deps) == 1 ? "y" : "ies"
+                    print(iostr, "\n  ",
+                        color_string(string(length(precomperr_deps)), Base.warn_color()),
+                        " dependenc$(pluralpc) failed but may be precompilable after restarting julia"
+                    )
+                end
+                if internal_call && !isempty(failed_deps)
+                    plural1 = length(failed_deps) == 1 ? "y" : "ies"
+                    plural2 = length(failed_deps) == 1 ? "" : "s"
+                    print(iostr, "\n  ", color_string("$(length(failed_deps))", Base.error_color()), " dependenc$(plural1) errored. ")
+                    print(iostr, "To see a full report either run `import Pkg; Pkg.precompile()` or load the package$(plural2)")
+                end
             end
-            if n_loaded > 0
-                plural1 = n_loaded == 1 ? "y" : "ies"
-                plural2 = n_loaded == 1 ? "a different version is" : "different versions are"
-                plural3 = n_loaded == 1 ? "" : "s"
-                print(iostr, "\n  ",
-                    color_string(string(n_loaded), Base.warn_color()),
-                    " dependenc$(plural1) precompiled but $(plural2) currently loaded. Restart julia to access the new version$(plural3)"
-                )
-            end
-            if !isempty(precomperr_deps)
-                pluralpc = length(precomperr_deps) == 1 ? "y" : "ies"
-                print(iostr, "\n  ",
-                    color_string(string(length(precomperr_deps)), Base.warn_color()),
-                    " dependenc$(pluralpc) failed but may be precompilable after restarting julia"
-                )
-            end
-            if internal_call && !isempty(failed_deps)
-                plural1 = length(failed_deps) == 1 ? "y" : "ies"
-                plural2 = length(failed_deps) == 1 ? "" : "s"
-                print(iostr, "\n  ", color_string("$(length(failed_deps))", Base.error_color()), " dependenc$(plural1) errored. ")
-                print(iostr, "To see a full report either run `import Pkg; Pkg.precompile()` or load the package$(plural2)")
+            # show any stderr output, even if Pkg.precompile has been interrupted (quick_exit=true), given user may be
+            # interrupting a hanging precompile job with stderr output. julia#48371
+            if !isempty(stderr_outputs)
+                plural1 = length(stderr_outputs) == 1 ? "y" : "ies"
+                plural2 = length(stderr_outputs) == 1 ? "" : "s"
+                print(iostr, "\n  ", color_string("$(length(stderr_outputs))", Base.warn_color()), " dependenc$(plural1) had warnings during precompilation:")
+                for (pkgid, err) in stderr_outputs
+                    err = join(split(err, "\n"), color_string("\n│  ", Base.warn_color()))
+                    print(iostr, color_string("\n┌ ", Base.warn_color()), pkgid, color_string("\n│  ", Base.warn_color()), err, color_string("\n└  ", Base.warn_color()))
+                end
             end
         end
         let str=str
@@ -1489,6 +1536,7 @@ function precompile(ctx::Context, pkgs::Vector{String}=String[]; internal_call::
                 println(io, str)
             end
         end
+        quick_exit && return
         if !internal_call
             err_str = ""
             n_direct_errs = 0
@@ -1515,9 +1563,13 @@ function save_precompile_state()
     path = Operations.pkg_scratchpath()
     for (prefix, store) in (("suspend_cache_", pkgs_precompile_suspended), ("pending_cache_", pkgs_precompile_pending))
         fpath = joinpath(path, string(prefix, hash(string(Base.active_project(), Base.VERSION))))
-        mkpath(path); Base.Filesystem.rm(fpath, force=true)
-        open(fpath, "w") do io
-            serialize(io, store)
+        if isempty(store)
+            Base.rm(fpath, force=true)
+        else
+            mkpath(path); Base.rm(fpath, force=true)
+            open(fpath, "w") do io
+                serialize(io, store)
+            end
         end
     end
     return nothing
@@ -1534,6 +1586,7 @@ function recall_precompile_state()
                     empty!(store)
                 end
             end
+            Base.rm(fpath, force=true)
         else
             empty!(store)
         end